facial_authentication/requirements.txt
absl-py==1.0.0
attrs==21.4.0
certifi==2021.10.8
charset-normalizer==2.0.10
cycler==0.11.0
fonttools==4.29.1
idna==3.3
kiwisolver==1.3.2
matplotlib==3.5.1
mediapipe==0.8.9.1
numpy==1.21.2
opencv-contrib-python==4.5.5.62
opencv-python==4.5.5.62
packaging==21.3
Pillow==8.3.2
protobuf==3.19.4
pyparsing==3.0.7
pyserial==3.5
python-dateutil==2.8.2
pywin32==303 # change this out
requests==2.27.1
six==1.16.0
ttkbootstrap==1.5.1
urllib3==1.26.8
websockets==10.0
